api/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
openai
langchain
python-dotenv
